HUM-882 Stop hbird services on reset
Since we are manually calling umount to unmount the volumes on reset,
the most common reason for error is that the services are running.  This
*should* work for most cases.

If this turns out to be not enough, we can take the next drastic step
and call out to `fuser` to kill any processes accessing the volumes
before unmounting.